rearranged algorithms
