+++ morpheus/algorithms/impl/dynamic_matrix/print_impl_serial.hpp
@@ -1,5 +1,5 @@
 /**
- * print.inl
+ * print_impl_serial.hpp
  *
  * EPCC, The University of Edinburgh
  *
@@ -21,52 +21,24 @@
  * limitations under the License.
  */
 
-#ifndef MORPHEUS_ALGORITHMS_IMPL_PRINT_INL
-#define MORPHEUS_ALGORITHMS_IMPL_PRINT_INL
+#ifndef MORPHEUS_ALGORITHMS_IMPL_DYNAMIC_MATRIX_PRINT_IMPL_SERIAL_HPP
+#define MORPHEUS_ALGORITHMS_IMPL_DYNAMIC_MATRIX_PRINT_IMPL_SERIAL_HPP
 
 #include <iostream>
 #include <iomanip>
 
-#include <morpheus/core/exceptions.hpp>
 #include <morpheus/containers/dynamic_matrix.hpp>
 #include <morpheus/containers/vector.hpp>
 
 namespace Morpheus {
+// forward decl
+template <typename Printable>
+void print(const Printable& p);
+
+template <typename Printable, typename Stream>
+void print(const Printable& p, Stream& s);
+
 namespace Impl {
-
-template <typename Printable, typename Stream>
-void print(const Printable& p, Stream& s, Morpheus::CooTag) {
-  using index_type = typename Printable::index_type;
-  s << p.name() << "<" << p.nrows() << ", " << p.ncols() << "> with "
-    << p.nnnz() << " entries\n";
-
-  for (index_type n = 0; n < p.nnnz(); n++) {
-    s << " " << std::setw(14) << p.row_indices[n];
-    s << " " << std::setw(14) << p.column_indices[n];
-    s << " " << std::setprecision(4) << std::setw(8) << "(" << p.values[n]
-      << ")\n";
-  }
-}
-
-template <typename Printable, typename Stream>
-void print(const Printable& p, Stream& s, Morpheus::Impl::SparseMatTag) {
-  s << p.name() << "<" << p.nrows() << ", " << p.ncols() << "> with "
-    << p.nnnz() << " entries\n";
-  Morpheus::NotImplementedException(
-      "void print(const Printable& p, Stream& s, "
-      "Morpheus::Impl::SparseMatTag)");
-}
-
-template <typename Printable, typename Stream>
-void print(const Printable& p, Stream& s, Morpheus::DenseVectorTag) {
-  using index_type = typename Printable::size_type;
-  s << p.name() << "<" << p.size() << "> with " << p.size() << " entries\n";
-
-  for (index_type n = 0; n < p.size(); n++) {
-    s << " " << std::setw(14) << n;
-    s << " " << std::setprecision(4) << std::setw(8) << "(" << p[n] << ")\n";
-  }
-}
 
 struct print_fn {
   using result_type = void;
@@ -92,17 +64,6 @@ void print(const Printable& p, Stream& s, Morpheus::DynamicTag) {
 }
 
 }  // namespace Impl
-
-template <typename Printable>
-void print(const Printable& p) {
-  Morpheus::print(p, std::cout);
-}
-
-template <typename Printable, typename Stream>
-void print(const Printable& p, Stream& s) {
-  Morpheus::Impl::print(p, s, typename Printable::tag());
-}
-
 }  // namespace Morpheus
 
-#endif  // MORPHEUS_ALGORITHMS_IMPL_PRINT_INL
+#endif  // MORPHEUS_ALGORITHMS_IMPL_DYNAMIC_MATRIX_PRINT_IMPL_SERIAL_HPP
