+++ morpheus/algorithms/impl/print_impl.hpp
@@ -1,5 +1,5 @@
 /**
- * print.hpp
+ * print_impl.hpp
  *
  * EPCC, The University of Edinburgh
  *
@@ -21,23 +21,13 @@
  * limitations under the License.
  */
 
-#ifndef MORPHEUS_ALGORITHMS_PRINT_HPP
-#define MORPHEUS_ALGORITHMS_PRINT_HPP
+#ifndef MORPHEUS_ALGORITHMS_IMPL_PRINT_IMPL_HPP
+#define MORPHEUS_ALGORITHMS_IMPL_PRINT_IMPL_HPP
 
-#include <morpheus/algorithms/impl/print_impl.hpp>
+#include <morpheus/algorithms/impl/coo_matrix/print_impl_serial.hpp>
+#include <morpheus/algorithms/impl/csr_matrix/print_impl_serial.hpp>
+#include <morpheus/algorithms/impl/dia_matrix/print_impl_serial.hpp>
+#include <morpheus/algorithms/impl/dynamic_matrix/print_impl_serial.hpp>
+#include <morpheus/algorithms/impl/vector/print_impl_serial.hpp>
 
-namespace Morpheus {
-
-template <typename Printable, typename Stream>
-void print(const Printable& p, Stream& s) {
-  Morpheus::Impl::print(p, s, typename Printable::tag());
-}
-
-template <typename Printable>
-void print(const Printable& p) {
-  Morpheus::print(p, std::cout);
-}
-
-}  // namespace Morpheus
-
-#endif  // MORPHEUS_ALGORITHMS_PRINT_HPP
+#endif  // MORPHEUS_ALGORITHMS_IMPL_PRINT_IMPL_HPP
